Allow users to run `tric here` at the site root
Rather than restricting `tric` use cases to plugins _only_, this changeset allows for the potential of running tests at the site or theme level as well.

To do this, I needed to extend the docker-compose file (`tric-stack.yml`) with `tric-stack.site.yml` as the paths require some tweaks when setting the WP path directly.