Update npm and composer functions to use the new parallel approach
We had a couple of functions that were used during `tric init` that were _not_ parallel executions of the npm and composer commands. This changeset reorganizes the code a bit so the base commands use these functions.

Additionally, these npm and composer functions now accept a second parameter that holds an array of subdirectories to prompt for execution.