Prompt user to update if tric version mismatches current build version
This adds a new non-git tracked file called `.build-version` that tracks the version of `tric` when the containers were last updated. This way, upon updating `tric`, if there is a version mismatch, the user is prompted to update their containers.

The prompt is _not_ triggered if the user does not have a `.env.tric.run` file as it will assume the user is running commands for the first time.

**Additionally:** I've added the output of the npm error log if one gets generated (sanse the saveTree lines). It is that change that triggered me to want to have a prompt because I was stumped for 2 minutes or so before I remembered I needed to rebuild the container stack.

:movie_camera: http://p.tri.be/bYtE6L