Allow npm install on subdirectory use targets
The problem with `npm install` on plugin subdirectories was that the presence of a `.git/` directory was _required_ to perform the git commands that npm was executing. Because `common/` is a submodule, it doesn't have a `.git/` directory...so when we were mounting common as `/project`, the `.git/` dir was not found, even when traversing up the tree.

The solution was to—when running `tric use` against a submoduled subdirectory—set the parent directory as the volume and then `cd` into `common/` in the `docker-entrypoint.sh`.

:movie_camera: http://p.tri.be/7dw7hH