Add host-ip command.
When configuring `tric` on systems that do not support the
out-of-the-box `host.internal.docker` redirection, i.e. any Linux
system, then finding out the IP Address of the Docker host machine from
the perspective of the the containers can be a bit tricky and
unreliable.

The purpose of this command is to make it easy to find that IP Address
and script its usage by means of the `-q` (for "quiet") option.