fix(trick-stack.yml) set CC container version.
In the previous version, `0.5.31`, I had set the version of the
`lucatume/codeception` container to `latest`.
That would pull the new version based on PHP 7.4, not PHP 7.3.
The PHP version upgrade would break some vendor libraries not compatible
with PHP 7.4.
This commit fixes the version of the container to
`cc3.1.0-v1.1.1`, the `cc3.1.0` part will make it so the base container
is based on PHP version 7.4, and the fixed version makes sure `tric`
will not depend on the latest version, but on a very specific one.