Update for PHP 8
Can't pass null or false to preg_replace for the third param - it throws a deprecation message.
`Deprecated: preg_replace(): Passing null to parameter #3 ($subject) of type array|string is deprecated in slic/src/scaffold.php on line 123`

Set the default to an empty string and cast the return of file_get_contents() to a string (it can return false).