Set the composer-cache directory correctly
This ensures that there is always a `/composer-cache` directory in the container and defaults to a `.cache/` directory in the `slic` project, but allows for overriding elsewhere.

Previously, the mounted volumes and composer cache dir env variable were set poorly when `slic composer-cache set <dir>` _wasn't run_, causing `files/`, `vcs/`, and `repo/` directories to appear on the filesystem after running `slic composer install`.