build(slic-stack.yml) set the PLAYWRIGHT_BROWSERS_PATH env var
By default Playwright will place browsers used in testing in the
`$HOME/.cache/ms-playwright` directory. Since the installation of the
browser dependencies and the browser must happen as the `root` user,
this would place the browsers in the `/root/.cache/ms-playwright`
directory. This update fixes the path to the one the `slic` user would
naturally use, making it possible to install the browsers as the `root`
user in a location where the `slic` user will find and be able to use
them.