set MySQL to 5.5.62 on PHP 7.4
This change aligns `slic` with the minimum WordPress requirements of PHP
7.2 and MySQL 5.5.
As a concession to the fact that `slic` is a  StellarWP tool, the
minimum PHP version supported by it is PHP `7.4`.

This is treated like a minor breaking change since it's breaking by
default, but can be controller setting either the `SLIC_DB_IMAGE` or
`SLIC_DB_NO_MIN` environment variables.